project-inventory/src/main/java/com/enigma/arinda/model/items/ItemResponse.java
package com.enigma.arinda.model.items;

import com.enigma.arinda.model.UnitModel;

public class ItemResponse {

    private Integer id;
    private String name;
    private UnitModel unit;
    private Integer price;

    public ItemResponse(Integer id, String name, Integer price) {
        this.id = id;
        this.name = name;
        this.price = price;
    }

    public ItemResponse() {

    }

    public Integer getPrice() {
        return price;
    }

    public void setPrice(Integer price) {
        this.price = price;
    }

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public UnitModel getUnit() {
        return unit;
    }

    public void setUnit(UnitModel unit) {
        this.unit = unit;
    }
}
